原理上测试通过

+++ inc/include.h
@@ -40,7 +40,7 @@ extern char dllink[500];
 extern char songpoint[50];
 extern uint8_t flag_netpoll ;
 extern char *net_buff;
-extern uint16_t bypass_cnt;
+extern uint32_t bypass_cnt;
 extern ESP8266_DEF_PTR ESP8266_A;
 extern void net_init();
 extern int socket_request(unsigned char option);
